additional temp refacto's
