+++ app/services/bridge/banks/list_banks.rb
@@ -1,9 +1,8 @@
-class Bridge::Users::SingleBank
+class Bridge::Banks::ListBanks
   require 'rest-client'
 
-  def self.call(access_token)
-    # bank_id = 365
-    url = "https://sync.bankin.com/v2/banks/#{bank_id}"
+def self.call(access_token)
+    url = "https://sync.bankin.com/v2/banks"
     params = {
       client_id: ENV["bridge_client_id"],
       client_secret: ENV["bridge_client_secret"]
@@ -16,18 +15,3 @@ class Bridge::Users::SingleBank
     JSON.parse(response.body)["resources"]
   end
 end
-
-# def self.call(access_token)
-#     url = "https://sync.bankin.com/v2/banks"
-#     params = {
-#       client_id: ENV["bridge_client_id"],
-#       client_secret: ENV["bridge_client_secret"]
-#     }
-#     response = RestClient.get(url, {
-#       params: params,
-#       "Bankin-Version": "2018-06-15",
-#       "Authorization": "Bearer #{access_token}"
-#     })
-#     JSON.parse(response.body)["resources"]
-#   end
-# end
