+++ app/services/bridge/categories/list_all_categories.rb
@@ -1,17 +1,17 @@
-class Bridge::Accounts::SingleAccount
+class Bridge::Categories::ListAllCategories
   require 'rest-client'
 
-  def self.call(account_id, access_token)
-    # account_id = ???
-    url = "https://sync.bankin.com/v2/accounts/#{account_id}?client_id=#{ENV["bridge_client_id"]}&client_secret=#{ENV["bridge_client_secret"]}"
+  def self.call
+    url = "https://sync.bankin.com/v2/categories?limit=500"
     params = {
+      limit: 500,
       client_id: ENV["bridge_client_id"],
       client_secret: ENV["bridge_client_secret"]
     }
     response = RestClient.get(url, {
       params: params,
       "Bankin-Version": "2018-06-15",
-      "Authorization": "Bearer #{access_token}"
+      "Accept-Language": "FR"
     })
     JSON.parse(response.body)["resources"]
   end
