+++ app/services/bridge/items/list_items.rb
@@ -1,9 +1,8 @@
-class Bridge::Accounts::SingleAccount
+class Bridge::Items::ListItems
   require 'rest-client'
 
-  def self.call(account_id, access_token)
-    # account_id = ???
-    url = "https://sync.bankin.com/v2/accounts/#{account_id}?client_id=#{ENV["bridge_client_id"]}&client_secret=#{ENV["bridge_client_secret"]}"
+  def self.call(access_token)
+    url = "https://sync.bankin.com/v2/items"
     params = {
       client_id: ENV["bridge_client_id"],
       client_secret: ENV["bridge_client_secret"]
@@ -16,4 +15,3 @@ class Bridge::Accounts::SingleAccount
     JSON.parse(response.body)["resources"]
   end
 end
-
